# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Enerji Yönetim Sistemi API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /weather:
    get:
      operationId: listWeather
      tags: [weather]
      summary: HDD/CDD verileri listesi
      parameters:
        - name: year
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WeatherRecord"
    post:
      operationId: fetchWeatherData
      tags: [weather]
      summary: Meteoroloji API'sinden HDD/CDD çek
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WeatherFetchRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WeatherRecord"

  # ── ANALYSIS ─────────────────────────────────────────────
...
components:
  schemas:
...
    WeatherRecord:
      type: object
      required: [id, year, month, hdd, cdd, location]
      properties:
        id:
          type: integer
        year:
          type: integer
        month:
          type: integer
        hdd:
          type: number
        cdd:
          type: number
        location:
          type: string
        avgTemp:
          type: ["number", "null"]
        createdAt:
          type: string
    WeatherFetchRequest:
      type: object
      required: [location, year]
      properties:
        location:
          type: string
        year:
          type: integer

    # ── Regression / Performance ─────────────────────────────
...